Move version check to repository
