+++ crates/repository/Cargo.toml
@@ -8,13 +8,15 @@ homepage = "https://github.com/hulks/hulk"
 [dependencies]
 color-eyre = { workspace = true }
 constants = { workspace = true }
-glob = { workspace = true }
 futures-util = { workspace = true }
+glob = { workspace = true }
 home = { workspace = true }
 itertools = { workspace = true }
 parameters = { workspace = true }
+semver = { workspace = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
 spl_network_messages = { workspace = true }
 tempfile = { workspace = true }
 tokio = { workspace = true }
+toml = { workspace = true }
