--- crates/versions/Cargo.toml
@@ -1,12 +1,0 @@
-[package]
-name = "versions"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-color-eyre = { workspace = true }
-semver = { workspace = true }
-serde = { workspace = true }
-toml = { workspace = true }
